backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-dotenv==1.0.0
groq==0.36.0
openai==1.12.0
redis==5.0.1
python-socketio==5.11.0
# eventlet==0.33.3  # Not compatible with Python 3.12, using threading mode instead
Werkzeug==3.0.1
requests>=2.28.0
coqui-tts>=0.0.1

